moved lp provider, bidder accounts to test runner
