+++ scripts/integrationTests/smokeTest1/auctionOpen.ts
@@ -1,6 +1,5 @@
 import assert from "assert";
 import { DepositArgs, WithdrawArgs } from "../../utils/facades/types";
-import { getLiquidityProviderAccounts } from "../../utils/helpers/accounts";
 import { TestRunner } from "../../utils/facades/TestRunner";
 import { LibraryError } from "starknet";
 
@@ -14,9 +13,10 @@ export const smokeTest = async ({
   getLPUnlockedBalanceAll,
   depositAll,
   withdrawAll,
-  getBalancesAll
+  getBalancesAll,
+  getLiquidityProviderAccounts
 }: TestRunner) => {
-  const liquidityProviderAccounts = getLiquidityProviderAccounts(provider, 2);
+  const liquidityProviderAccounts = getLiquidityProviderAccounts(2);
 
   //Approve A for depositing
   await eth.approval({
