+++ scripts/utils/facades/RoundSimulator.ts
@@ -11,10 +11,6 @@ import { TestRunner } from "./TestRunner";
 import { getOptionRoundContract, getOptionRoundFacade } from "../helpers/setup";
 import { optionRoundABI } from "../../abi";
 import { OptionRoundFacade } from "./optionRoundFacade";
-import {
-  getLiquidityProviderAccounts,
-  getOptionBidderAccounts,
-} from "../helpers/accounts";
 
 export type SimulationSheet = {
   liquidityProviders: Array<number>;
@@ -56,8 +52,8 @@ export class RoundSimulator {
   ) {
     this.testRunner = testRunner;
     this.optionRoundFacade = new OptionRoundFacade(optionRoundContract);
-    this.lpAccounts = getLiquidityProviderAccounts(testRunner.provider, 5);
-    this.bidderAccounts = getOptionBidderAccounts(testRunner.provider, 5);
+    this.lpAccounts = testRunner.getLiquidityProviderAccounts(5);
+    this.bidderAccounts = testRunner.getOptionBidderAccounts(5);
   }
 
   async simulateRound(params: SimulationParameters) {
